+++ src/spectrochempy/examples/core/e_project/plot_project.py
@@ -68,49 +68,6 @@ newproj
 
 
 # %%
-# Now we add a script to the original proj
-
-script_source = """
-set_loglevel(INFO)
-info_('samples contained in the project are:%s'%proj.projects_names)
-"""
-
-proj["print_info"] = scp.Script("print_info", script_source)
-proj
-
-# %%
-# save but do not change the original data
-
-proj.save(overwrite_data=False)
-
-# %%
-# RELOAD it
-
-newproj = scp.Project.load("project_1")
-newproj
-
-# %%
-# Execute a script
-
-scp.run_script(newproj.print_info)
-
-# %%
-# Another way to do the same thing is ith the following syntax (which may
-# seem simpler
-
-newproj.print_info()
-
-# %%
-# Finally lets use a more useful script
-script_source_2 = """
-proj.A350.TG.plot_scatter(title='my scatter plot');
-#show()
-"""
-proj["tgscatter"] = scp.Script("tgscatter", script_source_2)
-
-proj.tgscatter()
-
-# %%
 # This ends the example ! The following line can be uncommented if no plot shows when
 # running the .py script with python